pyproject.toml
[tool.poetry]
name = "flib"
version = "0.1.0"
description = ""
authors = ["acordier <acordier@fonctionlabs.com>",
           "mcordier <mcordier@fonctionlabs.com>",
           "jvassalo <jvassalo@fonctionlabs.com>"]
readme = "README.md"
packages = [{include = "flib"}]

[tool.poetry.dependencies]
python = "^3.10"
qdrant-client = "^1.9.0"
openai = "^1.23.6" # TODO: not sure this includes openai-clip
pillow="^10.3.0"
ollama = "^0.2.0"
boto3 = "^1.34.94"
joblib = "^1.4.2"

# Optional dependencies below
segment-anything = { version = "^1.0", optional = true }
segment-anything-fast = { version = "^0.1.2", optional = true }
opencv-python = { version="^4.9.0.80", optional = true }
matplotlib = { version = "^3.8.4", optional = true }
torch = { version = "^2.2.2", optional = true }
torchvision = { version = "^0.17.2", optional = true }

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"

# Define optional dependencies here
pytest = "^8.3.3"
[tool.poetry.extras]
extra-clip = ["torch", "torchvision"]
extra-sam = ["segment-anything", "segment-anything-fast", "torch", "torchvision", "opencv-python", "matplotlib"]

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
